fix: show loader at the siderbar while preparing to load the data from either local storage or dataabase
