+++ .changeset/purple-mugs-hammer.md
@@ -1,0 +1,5 @@
+---
+"geohub": patch
+---
+
+fix: show loader at the siderbar while preparing to load the data from either local storage or dataabase
