+++ sites/geohub/src/components/pages/map/Map.svelte
@@ -4,14 +4,22 @@
 	import MapQueryInfoControl from '$components/pages/map/plugins/MapQueryInfoControl.svelte';
 	import StyleShareControl from '$components/pages/map/plugins/StyleShareControl.svelte';
 	import { AdminControlOptions, MapStyles, TourOptions, attribution } from '$lib/config/AppConfig';
-	import { fromLocalStorage, getSpriteImageList, storageKeys, toLocalStorage } from '$lib/helper';
+	import {
+		fromLocalStorage,
+		getSpriteImageList,
+		isStyleChanged,
+		storageKeys,
+		toLocalStorage
+	} from '$lib/helper';
 	import type { Layer } from '$lib/types';
 	import {
+		MAPSTORE_CONTEXT_KEY,
+		PAGE_DATA_LOADING_CONTEXT_KEY,
+		SPRITEIMAGE_CONTEXT_KEY,
 		layerList as layerListStore,
 		type MapStore,
-		MAPSTORE_CONTEXT_KEY,
-		type SpriteImageStore,
-		SPRITEIMAGE_CONTEXT_KEY
+		type PageDataLoadingStore,
+		type SpriteImageStore
 	} from '$stores';
 	import MaplibreCgazAdminControl from '@undp-data/cgaz-admin-tool';
 	import StyleSwicher from '@undp-data/style-switcher';
@@ -25,32 +33,169 @@
 		ScaleControl,
 		TerrainControl,
 		type MapOptions,
+		type StyleSpecification,
 		type TerrainSpecification
 	} from 'maplibre-gl';
 	import { getContext, onMount } from 'svelte';
 
 	const map: MapStore = getContext(MAPSTORE_CONTEXT_KEY);
 	const spriteImageList: SpriteImageStore = getContext(SPRITEIMAGE_CONTEXT_KEY);
+	const pageDataLoadingStore: PageDataLoadingStore = getContext(PAGE_DATA_LOADING_CONTEXT_KEY);
 
 	let tourOptions: TourGuideOptions;
 	let tourLocalStorageKey = `geohub-map-${$page.url.host}`;
 
 	let container: HTMLDivElement;
-	// let map: Map;
-	export let mapOptions: MapOptions;
-	export let layerList: Layer[];
+
 	export let defaultStyle: string = MapStyles[0].title;
 
 	const layerListStorageKey = storageKeys.layerList($page.url.host);
 	const mapStyleStorageKey = storageKeys.mapStyle($page.url.host);
+	const mapStyleIdStorageKey = storageKeys.mapStyleId($page.url.host);
 	const initialLayerList: Layer[] | null = fromLocalStorage(layerListStorageKey, null);
+	const initiaMapStyle: StyleSpecification | null = fromLocalStorage(mapStyleStorageKey, null);
+	const initiaMapStyleId: string = fromLocalStorage(mapStyleIdStorageKey, null)?.toString();
 
 	const terrainOptions: TerrainSpecification = {
 		source: 'terrarium',
 		exaggeration: 1
 	};
 
+	let mapOptions: MapOptions = {
+		container: undefined,
+		style: $page.data.defaultStyle,
+		center: [0, 0],
+		zoom: 3,
+		hash: true,
+		attributionControl: false
+	};
+
 	onMount(() => {
+		retrieveExistingMapStyle().then(mapInitialise);
+	});
+
+	const retrieveExistingMapStyle = async () => {
+		const style = $page.data.style;
+		if (style) {
+			// /map/{id} page
+
+			// if style query param in URL
+			if (`${initiaMapStyleId}` === `${style.id}`) {
+				// If style id in local storage is the same with style query param
+				if (isStyleChanged(initiaMapStyle, style.style)) {
+					// restore from local storage
+
+					// to make sure dataset links exist, otherwise restore from database
+					let linksNotExist = true;
+					initialLayerList.forEach((l) => {
+						if (l.dataset.properties.links) {
+							linksNotExist = false;
+						}
+					});
+					if (!linksNotExist) {
+						const nonExistLayers: Layer[] = [];
+						for (const l of initialLayerList) {
+							const id = l.dataset.properties.id;
+							const stacType = l.dataset.properties.tags.find((t) => t.key === 'stacType')?.value;
+							if (['cog', 'mosaicjson'].includes(stacType)) continue;
+							const datasetUrl = `${$page.url.origin}/api/datasets/${id}`;
+							const res = await fetch(datasetUrl);
+							if (res.ok) {
+								l.dataset = await res.json();
+							} else {
+								nonExistLayers.push(l);
+							}
+						}
+						// only accept if dataset metadata is fetched
+						if (nonExistLayers.length > 0) {
+							nonExistLayers.forEach((layer) => {
+								initialLayerList.splice(initialLayerList.findIndex((l) => l.id === layer.id));
+								initiaMapStyle.layers.splice(
+									initiaMapStyle.layers.findIndex((l) => l.id === layer.id)
+								);
+							});
+						}
+						restoreStyle(initiaMapStyle, initialLayerList);
+						toLocalStorage(mapStyleStorageKey, initiaMapStyle);
+						toLocalStorage(layerListStorageKey, initialLayerList);
+					} else {
+						restoreStyle(style.style, style.layers);
+					}
+				} else {
+					initialLayerList?.forEach((l) => {
+						const l2 = style.layers.find((x) => x.id === l.id);
+						if (l2?.activeTab !== l.activeTab) {
+							l2.activeTab = l.activeTab;
+						}
+					});
+
+					// restore from database
+					restoreStyle(style.style, style.layers);
+				}
+			} else {
+				// style ID is different from query param
+				restoreStyle(style.style, style.layers);
+				toLocalStorage(mapStyleIdStorageKey, style.id);
+			}
+		} else {
+			// /map page
+
+			if (initiaMapStyle && initialLayerList && initialLayerList.length > 0) {
+				let existAllLayers = true;
+				initialLayerList.forEach((l) => {
+					if (!initiaMapStyle.layers.find((ml) => ml.id === l.id)) {
+						existAllLayers = false;
+					}
+				});
+				if (existAllLayers) {
+					// restore from local storage
+
+					// to make sure dataset links exist, otherwise remove all local storage
+					let linksNotExist = true;
+					initialLayerList.forEach((l) => {
+						if (l.dataset.properties.links) {
+							linksNotExist = false;
+						}
+					});
+					if (!linksNotExist) {
+						const nonExistLayers: Layer[] = [];
+						for (const l of initialLayerList) {
+							const id = l.dataset.properties.id;
+							const stacType = l.dataset.properties.tags.find((t) => t.key === 'stacType')?.value;
+							if (['cog', 'mosaicjson'].includes(stacType)) continue;
+							const datasetUrl = `${$page.url.origin}/api/datasets/${id}`;
+							const res = await fetch(datasetUrl);
+							if (res.ok) {
+								l.dataset = await res.json();
+							} else {
+								nonExistLayers.push(l);
+							}
+						}
+						// only accept if dataset metadata is fetched
+						if (nonExistLayers.length > 0) {
+							nonExistLayers.forEach((layer) => {
+								initialLayerList.splice(initialLayerList.findIndex((l) => l.id === layer.id));
+								initiaMapStyle.layers.splice(
+									initiaMapStyle.layers.findIndex((l) => l.id === layer.id)
+								);
+							});
+						}
+						restoreStyle(initiaMapStyle, initialLayerList);
+						toLocalStorage(mapStyleStorageKey, initiaMapStyle);
+						toLocalStorage(layerListStorageKey, initialLayerList);
+					} else {
+						toLocalStorage(layerListStorageKey, []);
+						toLocalStorage(mapStyleStorageKey, $page.data.defaultStyle);
+					}
+				} else {
+					toLocalStorage(layerListStorageKey, []);
+					toLocalStorage(mapStyleStorageKey, $page.data.defaultStyle);
+				}
+			}
+		}
+	};
+
+	const mapInitialise = () => {
 		mapOptions.container = container;
 		$map = new Map(mapOptions);
 
@@ -122,14 +267,9 @@
 				'top-right'
 			);
 
-			$layerListStore = [...layerList];
-
 			layerListStore.subscribe((value) => {
-				const storageValue = value
-					? value
-					: initialLayerList && initialLayerList.length > 0
-					? initialLayerList
-					: null;
+				const layerList: Layer[] | null = fromLocalStorage(layerListStorageKey, []);
+				const storageValue = value ? value : layerList && layerList.length > 0 ? layerList : null;
 				toLocalStorage(layerListStorageKey, storageValue);
 			});
 
@@ -141,8 +281,33 @@
 				let storageValue = $map.getStyle();
 				toLocalStorage(mapStyleStorageKey, storageValue);
 			});
+
+			$pageDataLoadingStore = false;
 		});
-	});
+	};
+
+	const restoreStyle = (newStyle: StyleSpecification, newLayerList: Layer[]) => {
+		mapOptions.style = newStyle;
+
+		if (newStyle.center) {
+			mapOptions.center = [newStyle.center[0], newStyle.center[1]];
+		}
+		if (newStyle.zoom) {
+			mapOptions.zoom = newStyle.zoom;
+		}
+
+		if (newStyle.bearing) {
+			mapOptions.bearing = newStyle.bearing;
+		}
+
+		if (newStyle.pitch) {
+			mapOptions.pitch = newStyle.pitch;
+		}
+
+		$layerListStore = [...newLayerList];
+		toLocalStorage(layerListStorageKey, newLayerList);
+		toLocalStorage(mapStyleStorageKey, newStyle);
+	};
 </script>
 
 <div bind:this={container} class="map" />
