+++ sites/geohub/src/lib/helper/getLayerStyle.ts
@@ -6,6 +6,6 @@ export const getLayerStyle = (
 	layerId: string
 ): RasterLayerSpecification | VectorLayerSpecification => {
 	const style = map.getStyle();
-	const layer = style.layers.find((l) => l.id === layerId);
+	const layer = style?.layers?.find((l) => l.id === layerId);
 	return layer as RasterLayerSpecification | VectorLayerSpecification;
 };
